Uploading images/content, deleting unused directories, renaming existing directories
