--- src/assets/img/.gitkeep
@@ -1,1 +1,0 @@
-# You can delete this file. It's just here to make Git happy.
