+++ src/index.html
@@ -1,0 +1,117 @@
+<!doctype html>
+<html class="no-js" lang="en">
+<head>
+	<meta charset="utf-8" />
+	<meta name="viewport" content="width=device-width, initial-scale=1">
+	<meta http-equiv="X-UA-Compatible" content="IE=edge" />
+	<title>August Ash Web Development - Minneapolis, MN</title>
+	<script src="https://cdnjs.cloudflare.com/ajax/libs/modernizr/2.8.3/modernizr.min.js" integrity="sha256-0rguYS0qgS6L4qVzANq4kjxPLtvnp5nn2nB5G1lWRv4=" crossorigin="anonymous"></script>
+	<link rel="stylesheet" href="/css/app.css" />
+	<link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Lato:normal,bold">
+	<link rel="stylesheet" type="text/css" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.3.0/css/font-awesome.min.css" />
+	<link href="img/favicon.ico" rel="shortcut icon" type="image/x-icon" />
+</head>
+<body>
+<div id="top">
+<header id="navContainer">
+	<div class="row collapse">
+		<section class="small-8 small-offset-4 large-6 large-offset-0 columns text-center logoArea">
+			<a href="#"><h1><img src="img/logo.svg" alt="August Ash - Minneapolis, Minnesota"></h1></a>
+		</section>
+		<a class="sr-only" href="#mainContent">Skip to main content</a>
+		<!-- START: Nav -->
+		<nav class="small-16 large-10 columns align-self-bottom">
+			<div class="title-bar" data-responsive-toggle="mainNav" data-hide-for="large">
+				<button class="menu-icon" type="button" data-toggle></button>
+				<div class="title-bar-title">Navigation</div>
+			</div>
+			<ul id="mainNav" class="menu large-horizontal medium-vertical" data-responsive-menu="accordion large-dropdown medium-accordion">
+				<li><a href="#">Who We Are</a>
+					<ul class="menu">
+						<li>Lorem ipsum dolor</li>
+					</ul>
+				</li>
+				<li><a href="#">What We Do</a></li>
+				<li><a href="#">Our Work</a></li>
+				<li><a href="#">Our Blog</a></li>
+				<li><a href="#">Careers</a></li>
+			</ul>
+		</nav>
+		<!-- END: /Nav -->
+	</div>
+</header>
+<!-- START: Slideshow -->
+<div id="theSlideshow">
+	<div class="orbit" role="region" aria-label="Slideshow" data-orbit>
+		<button class="orbit-previous"><span class="show-for-sr">Previous Slide</span>&#9664;&#xFE0E;</button>
+		<button class="orbit-next"><span class="show-for-sr">Next Slide</span>&#9654;&#xFE0E;</button>
+		<ul class="orbit-container">
+			<li class="is-active orbit-slide">
+				<img class="orbit-image" src="http://placehold.it/1200x400&amp;text=wuhn" alt="Placeholder 1">
+			</li>
+			<li class="orbit-slide">
+				<img class="orbit-image" src="http://placehold.it/1200x400&amp;text=tew" alt="Placeholder 2">
+			</li>
+			<li class="orbit-slide">
+				<img class="orbit-image" src="http://placehold.it/1200x400&amp;text=tree" alt="Placeholder 3">
+			</li>
+			<li class="orbit-slide">
+				<img class="orbit-image" src="http://placehold.it/1200x400&amp;text=foar" alt="Placeholder 4">
+			</li>
+		</ul>
+	</div>
+</div>
+<!-- END: /Slideshow -->
+<!-- START: Main Content -->
+<div id="mainContent">
+	<div id="contentBox" class="row"  role="main">
+		<section class="small-16 columns">
+			<h2 class="small-text-center">Aron</h2>
+			<p><strong>Flannel disrupt fashion axe deep v etsy.</strong> Pork belly locavore farm-to-table skateboard fap. Plaid 90's kale chips cred put a bird on it hashtag. Tote bag knausgaard twee, health goth VHS sartorial viral food truck typewriter. Tote bag listicle authentic echo park actually keffiyeh. Bespoke occupy direct trade cronut fingerstache. Meditation leggings pitchfork, craft beer VHS tattooed meh asymmetrical bushwick microdosing banjo tacos.</p>
+			<p>Knausgaard tousled 90's, chambray twee hoodie vice XOXO portland hella. Taxidermy mustache cornhole, health goth pitchfork church-key seitan art party man bun swag YOLO PBR&amp;B offal. Salvia 8-bit blog keffiyeh, farm-to-table drinking vinegar seitan. Ethical tofu hoodie next level. <a href="#">PBR&amp;B keffiyeh gluten-free schlitz</a>, pinterest readymade DIY sartorial. Ramps aesthetic actually salvia cold-pressed. Semiotics aesthetic cray, celiac chambray craft beer kale chips pour-over 8-bit hashtag kitsch austin.</p>
+			<p>Keytar salvia seitan pour-over banjo. Direct trade austin plaid banjo selvage, cliche blog retro chillwave. Pop-up slow-carb humblebrag brooklyn, williamsburg taxidermy ugh scenester. Health goth 8-bit meditation, affogato hella brunch pork belly austin sustainable. Biodiesel chia actually flexitarian selfies, umami meh etsy quinoa four loko yr thundercats flannel green juice. Pickled fap direct trade marfa iPhone neutra. Franzen ethical craft beer meggings, dreamcatcher normcore next level fashion axe selfies listicle tumblr.</p>
+		</section>
+	</div>
+</div>
+<!-- END: /Main Content -->
+<!-- START: Site Footer -->
+<footer id="siteFooter">
+	<div class="row" role="navigation">
+		<div class="small-16 columns footNav noPrint" role="navigation">
+			<ul class="menu">
+				<li><a href="#">Who We Are</a></li>
+				<li><a href="#">What We Do</a></li>
+				<li><a href="#">Our Work</a></li>
+				<li><a href="#">Our Blog</a></li>
+				<li><a href="#">Careers</a></li>
+			</ul>
+			<ul class="menu">
+				<li><a href="#">Contact Us</a></li>
+				<li><a href="#">Off-hour Support</a></li>
+				<li><a href="#">Submit Project</a></li>
+			</ul>
+			<section id="socialMedia" class="small-16 text-center noPrint">
+				<ul>
+					<li class="liLogo"><a href="#"><span class="fa fa-linkedin fa-2x"></span></a></li>
+					<li class="conLogo"><a href="#"><span class="fa fa-envelope fa-2x"></span></a></li>
+				</ul>
+			</section>
+			<p class="text-center">
+				<a href="tel:9528519400">952-851-9400</a><br />
+				2626 East 82nd Street, Suit 320 Bloomington, MN 55425<br />
+				&copy; 2016 August Ash
+			</p>
+		</div>
+	</div> <!-- .row -->
+</footer>
+<!-- END: /Site Footer -->
+</div>
+<script src="https://code.jquery.com/jquery-3.2.1.min.js" integrity="sha256-hwg4gsxgFZhOsEEamdOYGBf13FyQuiTwlAQgxVSNgt4=" crossorigin="anonymous"></script>
+<script type="text/javascript" src="js/app.js"></script>
+<script>
+$(document).ready(function(){
+	console.log('Foundation Core Version: ' + Foundation.version);
+});
+</script>
+</body>
+</html>
