+++ src/js/app.js
@@ -3,10 +3,10 @@ import whatInput from 'what-input';
 
 window.$ = $;
 
-import Foundation from 'foundation-sites';
+//import Foundation from 'foundation-sites';
 // If you want to pick and choose which modules to include, comment out the above and uncomment
 // the line below
-//import './lib/foundation-explicit-pieces';
+import './lib/foundation-explicit-pieces';
 
 
 $(document).foundation();
